% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binarizer.R
\name{binarizer}
\alias{binarizer}
\alias{binarizer_}
\alias{binarizer_}
\title{Binarize multiple columns at once}
\usage{
binarizer(data, ..., thresh = 0)

binarizer_(data, cols = NULL, thresh = 0)
}
\arguments{
\item{data}{Dataframe, tibble, etc.}

\item{..., cols}{Variables to include/exclude.

 ... :
 You can use same specifications as in dplyr's \link[dplyr]{select}.

 cols :
 character vector

 If missing, defaults to all non-grouping variables.}

\item{thresh}{Threshold (Numeric).}
}
\value{
Tibble where selected columns have been binarized.

Above thresh is 1; below or equal to thresh is 0.
}
\description{
Binarize multiple columns of a dataframe based on a given threshold.

 \strong{binarizer} is designed to work with \link[magrittr]{\%>\%} pipelines.

 \strong{binarizer_} is a standard evalution version.
}
\details{
Binarizes each specified column and converts to tibble.
}
\examples{
# Attach package
library(LRO.utilities)

# Create dataframe
df <- data.frame('a' = c(1,2,3,4,5,6,7),
                 'b' = c(2,3,4,5,6,7,8))

# First center both columns
centered_df <- scaler(df, scale = F)

# Binarizing multiple columns
binarizer(centered_df)
binarizer(centered_df, a, b)
binarizer(centered_df, 1:2)
binarizer(centered_df, c(a,b))

# Binarize 'a'
binarizer(centered_df, a)

# Binarize all but 'a'
binarizer(centered_df, -a)

## Standard evalutation versions

binarizer_(centered_df, cols = c('b'))

}
