% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/note_chunk.R
\name{insertNoteChunk}
\alias{insertNoteChunk}
\title{Insert note chunk.}
\usage{
insertNoteChunk()
}
\value{
Inserts \code{```{note, include=FALSE, eval=FALSE}\n\"\n\n\"\n```}
}
\description{
Addin for RStudio for inserting code chunks named "note" and with include=FALSE in R Markdown.

See \code{details} for setting up key command.
}
\details{
How to set up key command in RStudio:

After installing package.
Go to:

\code{Tools} >> \code{Addins} >> \code{Browse Addins} >> \code{Keyboard Shortcuts}.

Find \code{"Insert Note Chunk"} and press its field under \code{Shortcut}.

Press desired key command.

Press \code{Apply}.

Press \code{Execute}.
}
\author{
Ludvig Renbo Olsen, \email{r-pkgs@ludvigolsen.dk}
}
