% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_pipe.R
\name{insertPipe}
\alias{insertPipe}
\title{Insert pipe.}
\usage{
insertPipe()
}
\value{
Inserts \code{\%>\% }
}
\description{
Addin for RStudio for inserting pipe %>%.

See \code{details} for setting up key command.
}
\details{
How to set up key command in RStudio:

After installing package.
Go to:

\code{Tools} >> \code{Addins} >> \code{Browse Addins} >> \code{Keyboard Shortcuts}.

Find \code{"Insert Pipe"} and press its field under \code{Shortcut}.

Press desired key command.

Press \code{Apply}.

Press \code{Execute}.
}
\author{
Ludvig Renbo Olsen, \email{r-pkgs@ludvigolsen.dk}
}
