% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_col.R
\name{rename_col}
\alias{rename_col}
\title{Rename single column by name}
\usage{
rename_col(data, old_name, new_name)
}
\arguments{
\item{data}{Dataframe.}

\item{old_name}{Name of column to rename (Character).}

\item{new_name}{Desired name of column (Character).}
}
\value{
Dataframe with modified name of specified column.
}
\description{
Change name of a single column in a dataframe by old name instead of position.
}
