% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll_previous.R
\name{roll_previous}
\alias{roll_previous}
\title{Apply rolling functions to previous rows}
\usage{
roll_previous(v, width = 2, FUN = mean)
}
\arguments{
\item{v}{Vector.}

\item{width}{Number of previous row to apply function to.}

\item{FUN}{Function to apply.
Must be compatible with \link[zoo]{rollapply}.}
}
\description{
Wrapper for zoo::rollapply for
 applying a function with rolling windows and getting
 the result of the previous window.
 Appends NAs at start only.
}
\examples{

# Attach package
library(LRO.utilities)

# Create dataframe
df <- data.frame('round' = c(1,2,3,4,5,6,7,8,9,10),
                 'score' = c(5,3,4,7,6,5,2,7,8,6))

# For each row we find the mean score of the previous 2 rounds
df$mean_prev_score = roll_previous(df$score, width = 2, FUN = mean)
}
