% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxap.R
\name{maxap}
\alias{maxap}
\title{Finding maximum a posteriori (MAP) and HPDI from samples.}
\usage{
maxap(data, pars = c(), HPDI_prob = 0.89)
}
\arguments{
\item{data}{Dataframe with samples.
A column for each parameter}

\item{pars}{Names of parameters (Character).}

\item{HPDI_prob}{Probability for highest posterior density interval}
}
\description{
Given a dataframe with samples for parameters,
 find the MAP and HPDI.
}
