% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savage_dickey.R
\name{savage_dickey}
\alias{savage_dickey}
\title{Calculate Bayes Factor}
\usage{
savage_dickey(post, prior, Q, xlab = "Parameter values", ylab = "Density",
  print_plot = FALSE, return_plot = TRUE)
}
\arguments{
\item{post}{Samples from posterior distribution (Numeric)}

\item{prior}{Samples from prior distribution (Numeric)}

\item{Q}{Point on x-axis for calculating Bayes factor.}

\item{xlab}{Label for x-axis}

\item{ylab}{Label for y-axis}

\item{print_plot}{(Logical)}

\item{return_plot}{(Logical)}
}
\value{
List with ggplot2 object (optional), BF10 and BF01
}
\description{
Creates plot of prior and posterior distributions and calculates the
 Bayes factor at the point of interest (Q).
}
\details{
Bayes factors are calculated using the \link[polspline]{polspline} package.

 BF10 <- dlogspline(Q, logspline(post)) /
 dlogspline(Q,logspline(prior))

 BF01 <- dlogspline(Q, logspline(prior)) /
 dlogspline(Q,logspline(post))
}
\author{
Benjamin Hugh Zachariae

Ludvig Renbo Olsen
}
